pyproject.toml
[tool.poetry]
name = "unbosque-calendar"
version = "0.1.0"
description = ""
authors = ["wolf <wolfgangsalazarmora@gmail.com>"]
readme = "README.md"
packages = [{include = "unbosque_calendar"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
celery = "^5.3.1"
dj-database-url = "^2.0.0"
psycopg = {extras = ["binary"], version = "^3.1.9"}
whitenoise = "^6.5.0"
gunicorn = "^21.2.0"
oauthlib = "^3.2.2"
httpx = "^0.24.1"
result = "^0.13.1"
beautifulsoup4 = "^4.12.2"
google-api-python-client = "^2.96.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
